Changed penality calculation
The penality is now proportional between the number of days late and the number of days you were supposed to wait for the next study. If you initialy had to wait N days and you waited 2N, then the penality will be 10%. This will punish more severely early lateness, especially when you are only one day late for the first study of a subject.